Retrieve explorers functionality
